apps/scraper/requirements.txt
# KSU Research - Faculty Scraper
# Python dependencies

# Web scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
selenium>=4.15.0
playwright>=1.40.0

# Data processing
pandas>=2.1.0

# Database
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0

# Async HTTP
aiohttp>=3.9.0
asyncio>=3.4.3

# Rate limiting
ratelimit>=2.2.1

# Environment variables
python-dotenv>=1.0.0

# Progress bars
tqdm>=4.66.0

# Logging
loguru>=0.7.2

# Optional: Local embeddings (if using Ollama)
ollama>=0.1.0

# Optional: Google Scholar scraping
scholarly>=1.7.0
